Add mac os build workflow (#761)
* adding mac os build workflow

* fix vulkan install

* Vulkan paths added

* macos build fix

* Fix macOS bundle crash: copy Vulkan loader with correct soname

The binary links @rpath/libvulkan.1.dylib but the POST_BUILD step
was copying ${Vulkan_LIBRARIES} which resolves through symlinks to
libvulkan.dylib (without the .1 soname). dyld could not find the
library at launch, causing an immediate SIGABRT before main().

Resolve the real file and copy it explicitly as libvulkan.1.dylib.

Made-with: Cursor

* Use system Vulkan loader via rpath instead of bundling

Stop copying libvulkan.1.dylib into the app bundle. Instead, add
/usr/local/lib and /opt/homebrew/lib to the rpath so dyld finds
the Vulkan loader from the user's Homebrew or Vulkan SDK install.

Users must run `brew install vulkan-loader` for the app to launch.
MoltenVK remains optional — the app falls back to OpenGL without it.

Made-with: Cursor

---------

+++ src/view/src/rocprofvis_data_provider.cpp
@@ -881,7 +881,7 @@ DataProvider::ParseStreamData(rocprofvis_handle_t* stream_handle, StreamInfo& st
                                                   0, &stream_info.id);
         ROCPROFVIS_ASSERT(result == kRocProfVisResultSuccess);
         stream_info.name = GetString(stream_handle, kRPVControllerStreamName, 0);
-        size_t num_processors = 0;
+        uint64_t num_processors = 0;
         result = rocprofvis_controller_get_uint64(stream_handle, kRPVControllerStreamNumProcessors,
             0, &num_processors);
         ROCPROFVIS_ASSERT(result == kRocProfVisResultSuccess);
